Splited doc service and log service + fixed a problem when we couldn't edit changeDate for a log
