+++ src/main/java/docSharing/config/CustomWebSecurityConfigurerAdapter.java
@@ -16,7 +16,6 @@ public class CustomWebSecurityConfigurerAdapter {
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        System.out.println("in security filter chain");
         http.authorizeRequests().antMatchers("*").authenticated().and().httpBasic().and().csrf().disable();
 //        http.addFilterAfter(new AuthorizationFilter(), BasicAuthenticationFilter.class);
 //        http.addFilterAfter(new PermissionFilter(), BasicAuthenticationFilter.class);
