+++ src/main/java/docSharing/config/SecurityConfig.java
@@ -32,24 +32,4 @@ public class SecurityConfig {
                 .authenticated().and().csrf().disable().cors().configurationSource(request -> corsConfiguration);
 
     }
-
-//    @Bean
-//    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-//        System.out.println("MAYBE I AM HERE! SECURITY CONFIG");
-//        return http
-//                .cors(withDefaults()) // by default uses a Bean by the name of corsConfigurationSource
-//                .build();
-//    }
-//
-//    @Bean
-//    CorsConfigurationSource corsConfigurationSource() {
-//        CorsConfiguration configuration = new CorsConfiguration();
-//        configuration.setAllowedOrigins(Arrays.asList("http://localhost:3000"));
-//        configuration.setAllowedMethods(Arrays.asList("GET", "POST", "DELETE"));
-//        configuration.setAllowedHeaders(List.of("Authorization"));
-//        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-//        source.registerCorsConfiguration("/**", configuration);
-//        return source;
-//    }
-
 }
