+++ src/main/java/docSharing/entity/Log.java
@@ -26,7 +26,7 @@ public class Log {
     private Long id;
     @Column(name = "created_on", nullable = false, updatable = false)
     private LocalDateTime creationDate = LocalDateTime.now();
-    @Column(name = "edited_on", nullable = false, updatable = false)
+    @Column(name = "edited_on", nullable = false)
     private LocalDateTime lastEditDate;
     private Integer offset;
     private String action;
@@ -39,22 +39,21 @@ public class Log {
     private Document document;
 
 
-
-    public Log(User user, Document document, int offset, String data, String action) throws AccountNotFoundException {
+    public Log(User user, Document document, int offset, String data, String action, LocalDateTime lastEditDate) throws AccountNotFoundException {
         this.user = user;
         this.document = document;
         this.offset = offset;
         this.data = data;
         this.action = action;
+        this.lastEditDate = lastEditDate;
     }
 
     public static Log copy(Log log) {
         Log tempLog = new Log();
-        //tempLog.setUserId(log.getUserId());
         tempLog.setUser(log.getUser());
         tempLog.setOffset(log.getOffset());
         tempLog.setCreationDate(log.getCreationDate());
-        //tempLog.setDocumentId(log.getDocumentId());
+        tempLog.setLastEditDate((log.getLastEditDate()));
         tempLog.setDocument(log.getDocument());
         tempLog.setData(log.getData());
         tempLog.setAction(log.getAction());
