+++ src/main/java/docSharing/entity/SendLogsToDatabase.java
@@ -5,6 +5,8 @@ import docSharing.utils.debounce.Callback;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -12,28 +14,24 @@ import java.util.Map;
 public class SendLogsToDatabase implements Callback {
 
     private Map<Long, Map<Long, Log>> logsMap;
+    static List<Log> unsavedLogs;
 
-    public SendLogsToDatabase(Map<Long, Map<Long, Log>> logsMap) {
+
+    public SendLogsToDatabase(Map<Long, Map<Long, Log>> logsMap, List<Log> unsavedLogs) {
         this.logsMap = logsMap;
+        this.unsavedLogs = unsavedLogs;
     }
 
-//    @Autowired
-//    LogRepository logRepository;
-
     @Override
-    public void call(Log log, LogRepository logRepository) {
-
-        logsMap.get(log.getDocument().getId()).get(log.getUser().getId()).setLastEditDate(log.getLastEditDate());
-//       logRepository.save(logsMap.get(log.getDocument().getId()).get(log.getUser().getId()));
+    public void call(Log log) {
+        Log currentLog = logsMap.get(log.getDocument().getId()).get(log.getUser().getId());
+        currentLog.setLastEditDate(LocalDateTime.now().minusSeconds(3));
         // if the log's data is not empty or null, store it in the database
-        System.out.println("data: " + logsMap.get(log.getDocument().getId()).get(log.getUser().getId()).getData());
-        if (logsMap.get(log.getDocument().getId()).get(log.getUser().getId()).getData() != null || logsMap.get(log.getDocument().getId()).get(log.getUser().getId()).getData().length() > 0){
-            System.out.println("logsMap: " + logsMap);
-            logsMap.get(log.getDocument().getId()).remove(log.getUser().getId());
+        if (currentLog.getData() != null || currentLog.getData().length() > 0) {
+            // store the current log in a list that pushes all its data to the database one in a few seconds
+            unsavedLogs.add(currentLog);
+            // finally, remove the log from the map and clear its cache
+            logsMap.get(currentLog.getDocument().getId()).remove(currentLog.getUser().getId());
         }
-
-        // finally, remove the log from the map and clear its cache
-        //  logsMap.get(log.getDocument().getId()).remove(log.getUser().getId());
-
     }
 }
