+++ src/main/java/docSharing/utils/debounce/Debouncer.java
@@ -1,16 +1,14 @@
 package docSharing.utils.debounce;
 
 import docSharing.entity.Log;
-import docSharing.repository.LogRepository;
-
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-public class Debouncer <T> {
-    private final ScheduledExecutorService sched = Executors.newScheduledThreadPool(1);
-    private final ConcurrentHashMap<Long, TimerTask> delayedMap = new ConcurrentHashMap<Long, TimerTask>();
+public class Debouncer<T> {
+    private final ScheduledExecutorService schedule = Executors.newScheduledThreadPool(1);
+    private final ConcurrentHashMap<Long, TimerTask> delayedMap = new ConcurrentHashMap<>();
     private final Callback callback;
     private final int interval;
 
@@ -19,32 +17,30 @@ public class Debouncer <T> {
         this.interval = interval;
     }
 
-    public void call(Log log, LogRepository logRepository) {
-        TimerTask task = new TimerTask(log, logRepository);
+    public void call(Log log) {
+        TimerTask task = new TimerTask(log);
 
         TimerTask prev;
         do {
             prev = delayedMap.putIfAbsent(log.getUser().getId(), task);
             if (prev == null)
-                sched.schedule(task, interval, TimeUnit.MILLISECONDS);
+                schedule.schedule(task, interval, TimeUnit.MILLISECONDS);
         } while (prev != null && !prev.extend()); // Exit only if new task was added to map, or existing task was extended successfully
     }
 
     // STOP ALL RUNNING TASKS
-    public void terminate() {
-        sched.shutdownNow();
-    }
+//    public void terminate() {
+//        schedule.shutdownNow();
+//    }
 
     // The task that wakes up when the wait time elapses
     private class TimerTask implements Runnable {
         private final Log log;
-        private LogRepository logRepository;
         private long dueTime;
         private final Object lock = new Object();
 
-        public TimerTask(Log log, LogRepository logRepository) {
+        public TimerTask(Log log) {
             this.log = log;
-            this.logRepository=logRepository;
             extend();
         }
 
@@ -61,11 +57,11 @@ public class Debouncer <T> {
             synchronized (lock) {
                 long remaining = dueTime - System.currentTimeMillis();
                 if (remaining > 0) { // Re-schedule task
-                    sched.schedule(this, remaining, TimeUnit.MILLISECONDS);
+                    schedule.schedule(this, remaining, TimeUnit.MILLISECONDS);
                 } else { // Mark as terminated and invoke callback
                     dueTime = -1;
                     try {
-                        callback.call(log, logRepository);
+                        callback.call(log);
                     } finally {
                         delayedMap.remove(log.getUser().getId());
                     }
